// src/pricing/types/pricing.enums.ts
// LY:NOTE - In the real world, these prices would live within a database and be fetched from there. For this TH assignment I have hardcoded them here.
// Storing prices in a database allows for easier updating and management of prices without having to make changes to the code base.
export enum EPouchePrice {
  A = 55.5,
  B = 59.5,
  C = 62.75,
  D = 66.0,
  E = 69.0,
  F = 71.25,
}

export enum EPoucheType {
  A = 'A',
  B = 'B',
  C = 'C',
  D = 'D',
  E = 'E',
  F = 'F',
}